State's gameVariables are now doubles, POSITION_X/_Y/_Z gameVariables added (#127), objects' positions added to labels

+++ include/ViZDoomGame.h
@@ -82,7 +82,7 @@ namespace vizdoom {
         void addAvailableGameVariable(GameVariable var);
         void clearAvailableGameVariables();
         size_t getAvailableGameVariablesSize();
-        int getGameVariable(GameVariable variable);
+        double getGameVariable(GameVariable variable);
 
 
         /* GameArgs getters and setters */
